openroadm-driver/requirements.txt
ncclient 
lxml 
xmltodict 
python-dotenv
loguru